requirements.txt
flask
psutil
Pillow
opencv-python
